Create RequestHeadersFactory with Api and Default implementations
This simplifies the logic of the request classes
